ai-algorithm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
tensorflow==2.15.0
tensorflow-lite==2.15.0
numpy==1.24.3
Pillow==10.1.0
pandas==2.0.3
scikit-learn==1.3.0